tests/requirements-test.txt
pytest==7.4.0  # or the latest compatible version
pytest-asyncio==0.21.0 
requests==2.31.0  # or the latest compatible version
websocket-client==1.6.1  # or the latest compatible version
docker==6.1.3  # or the latest compatible version
aiohttp==3.8.5  # or the latest compatible version
ffmpeg-python==0.2.0  # or the latest compatible version
python-dotenv==1.0.0

# Any additional dependencies required by your tests

# device-manager/requirements.txt
#pyudev==0.24.3
#aiohttp==3.9.1
#obs-websocket-py==1.0.0
#zeroconf==0.128.0
#pyyaml==6.0.1
#v4l2==0.2
